// packages/lil-relayer/src/types/ethers-contracts/VRFCoordinator.ts
export declare namespace VRF {
  export type ProofStruct = {
    pk: [PromiseOrValue<BigNumberish>, PromiseOrValue<BigNumberish>];
    gamma: [PromiseOrValue<BigNumberish>, PromiseOrValue<BigNumberish>];
    c: PromiseOrValue<BigNumberish>;
    s: PromiseOrValue<BigNumberish>;
    seed: PromiseOrValue<BigNumberish>;
    uWitness: PromiseOrValue<string>;
    cGammaWitness: [PromiseOrValue<BigNumberish>, PromiseOrValue<BigNumberish>];
    sHashWitness: [PromiseOrValue<BigNumberish>, PromiseOrValue<BigNumberish>];
    zInv: PromiseOrValue<BigNumberish>;
  };

  export type ProofStructOutput = [
    [BigNumber, BigNumber],
    [BigNumber, BigNumber],
    BigNumber,
    BigNumber,
    BigNumber,
    string,
    [BigNumber, BigNumber],
    [BigNumber, BigNumber],
    BigNumber
  ] & {
    pk: [BigNumber, BigNumber];
    gamma: [BigNumber, BigNumber];
    c: BigNumber;
    s: BigNumber;
    seed: BigNumber;
    uWitness: string;
    cGammaWitness: [BigNumber, BigNumber];
    sHashWitness: [BigNumber, BigNumber];
    zInv: BigNumber;
  };

  export type RequestStruct = {
    sender: PromiseOrValue<string>;
    nonce: PromiseOrValue<BigNumberish>;
    blockNumber: PromiseOrValue<BigNumberish>;
    oracleId: PromiseOrValue<BytesLike>;
    requestConfirmations: PromiseOrValue<BigNumberish>;
    callbackGasLimit: PromiseOrValue<BigNumberish>;
    nbWords: PromiseOrValue<BigNumberish>;
    callbackSelector: PromiseOrValue<BytesLike>;
  };

  export type RequestStructOutput = [
    string,
    BigNumber,
    BigNumber,
    string,
    number,
    number,
    number,
    string
  ] & {
    sender: string;
    nonce: BigNumber;
    blockNumber: BigNumber;
    oracleId: string;
    requestConfirmations: number;
    callbackGasLimit: number;
    nbWords: number;
    callbackSelector: string;
  };
}